requirements.txt
# Core Dependencies
openai>=1.12.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
pypdf>=3.17.0

# NLP and Text Processing
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.6
tiktoken>=0.5.2
chromadb>=0.4.22

# ArXiv Integration
arxiv>=2.1.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
requests>=2.31.0
tenacity>=8.2.3
tqdm>=4.66.0

# Security
cryptography>=41.0.0

flask